+++ src/observer/sql/parser/parse_defs.h
@@ -191,7 +191,7 @@ void value_init_integer(Value *value, int v);
 void value_init_float(Value *value, float v);
 void value_init_string(Value *value, const char *v);
 void value_init_date(Value *value, const char *v);
-void long_value_init_date(Value *value, long v);
+void long_value_init_date(Value *value, int v);
 void value_destroy(Value *value);
 
 void condition_init(Condition *condition, CompOp comp, int left_is_attr, RelAttr *left_attr, Value *left_value,
