+++ src/observer/util/comparator.cpp
@@ -26,9 +26,9 @@ int compare_int(void *arg1, void *arg2)
 
 int compare_date(void *arg1, void *arg2)
 {
-  long v1 = *(long *)arg1;
-  long v2 = *(long *)arg2;
-  return int(v1 - v2);
+  int v1 = *(int *)arg1;
+  int v2 = *(int *)arg2;
+  return v1 - v2;
 }
 
 int compare_float(void *arg1, void *arg2)
